SimpleWhittedIntegrator has multiple sampling, progress-reporting, irradiance-scaling
Multiple Sampling:
Now each sample's resulting Spectrum is scaled in intensity by the total number of samples per pixel -- i.e., all samples for each pixel are averaged together. It's not a particularly accurate way of estimating with samples, but it's easy and "good enough" for now.

Progress-Reporting:
When each Sample-calculating thread is submitted, a percent-complete figure is computed. When it is 1 whole percentage higher than the last time, we print it to standard-output.

Irradiance-scaling:
We use the FresnelResult reflectance to scale the amount of specularly-reflected radiance that's added to the final outbound radiance. Now we also scale the surface irradiance (i.e., radiance contributed by light) by the same reflectance factor.